backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mietcheck-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "aiosmtplib==3.0.2",
    "jinja2==3.1.4",
    "reportlab==4.2.5",
    "httpx==0.27.2",
    "python-dateutil==2.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
